Improved the layer is in visible range flag updating, notably for the group layers
