+++ packages/geoview-core/src/geo/layer/gv-layers/abstract-base-layer.ts
@@ -2,11 +2,10 @@ import type BaseLayer from 'ol/layer/Base';
 import type { Projection as OLProjection } from 'ol/proj';
 
 import type { Extent } from '@/api/types/map-schema-types';
-import type { TypeLayerStatus } from '@/api/types/layer-schema-types';
+import type { TypeLayerStatus, EffectiveLayerScales } from '@/api/types/layer-schema-types';
 import type { EventDelegateBase } from '@/api/events/event-helper';
 import EventHelper from '@/api/events/event-helper';
 import type { ConfigBaseClass } from '@/api/config/validation-classes/config-base-class';
-import type { EffectiveLayerScales } from '@/geo/map/map-viewer';
 import { GVGroupLayer } from '@/geo/layer/gv-layers/gv-group-layer';
 
 /**
@@ -137,6 +136,49 @@ export abstract class AbstractBaseGVLayer {
 
     // If emitting
     if (emitZIndexChanged) this.#emitZIndexChanged({ zIndex });
+  }
+
+  /**
+   * Overridable method to get if the layer is in visible range.
+   *
+   * Compares against the OL layer's minResolution/maxResolution threshold.
+   * Uses inclusive boundaries on both ends (`<=`) to match ESRI/ArcGIS semantics where
+   * minScale and maxScale are both inclusive. This ensures a layer with maxScale=41999
+   * is visible at exactly scale 41999.
+   *
+   * @param currentResolution - Optional. The current map resolution in map units per pixel
+   * @param currentScale - Optional. The current map scale denominator (1:X)
+   * @param effectiveScales - Optional. Effective layer scales with buffer thresholds
+   * @returns True if the layer is in visible range
+   */
+  protected onIsInVisibleRange(
+    currentResolution: number | undefined,
+    currentScale?: number,
+    effectiveScales?: EffectiveLayerScales
+  ): boolean {
+    if (!currentResolution || Number.isNaN(currentResolution)) return false;
+    const minResolution = this.getOLLayer().getMinResolution(); // defaults to 0
+    const maxResolution = this.getOLLayer().getMaxResolution(); // defaults to Infinity
+    const inResolutionRange = currentResolution >= minResolution && currentResolution <= maxResolution;
+    if (!inResolutionRange) return false;
+
+    // If scale context is provided, enforce a strict "safe zone" by excluding values
+    // that fall inside configured visibility buffers around min/max scale thresholds.
+    if (currentScale && effectiveScales) {
+      const { maxScale, maxScaleTolerance, minScale, minScaleTolerance } = effectiveScales;
+
+      // Too close to zoom-in boundary (maxScale)
+      if (maxScale && maxScaleTolerance && currentScale >= maxScale && currentScale < maxScaleTolerance) {
+        return false;
+      }
+
+      // Too close to zoom-out boundary (minScale)
+      if (minScale && minScaleTolerance && currentScale > minScaleTolerance && currentScale <= minScale) {
+        return false;
+      }
+    }
+
+    return true;
   }
 
   // #endregion OVERRIDES
@@ -529,8 +571,7 @@ export abstract class AbstractBaseGVLayer {
   /**
    * Gets the in visible range value.
    *
-   * Compares against the OL layer's minResolution/maxResolution thresholds, which are
-   * set by updateLayerInVisibleRange from both config scale values and initialSettings zoom values.
+   * Compares against the OL layer's minResolution/maxResolution thresholds.
    * Uses inclusive boundaries on both ends (`<=`) to match ESRI/ArcGIS semantics where
    * minScale and maxScale are both inclusive. This ensures a layer with maxScale=41999
    * is visible at exactly scale 41999.
@@ -541,29 +582,8 @@ export abstract class AbstractBaseGVLayer {
    * @returns True if the layer is in visible range
    */
   isInVisibleRange(currentResolution: number | undefined, currentScale?: number, effectiveScales?: EffectiveLayerScales): boolean {
-    if (!currentResolution || Number.isNaN(currentResolution)) return false;
-    const minResolution = this.getOLLayer().getMinResolution(); // defaults to 0
-    const maxResolution = this.getOLLayer().getMaxResolution(); // defaults to Infinity
-    const inResolutionRange = currentResolution >= minResolution && currentResolution <= maxResolution;
-    if (!inResolutionRange) return false;
-
-    // If scale context is provided, enforce a strict "safe zone" by excluding values
-    // that fall inside configured visibility buffers around min/max scale thresholds.
-    if (currentScale && effectiveScales) {
-      const { maxScale, maxScaleTolerance, minScale, minScaleTolerance } = effectiveScales;
-
-      // Too close to zoom-in boundary (maxScale)
-      if (maxScale && maxScaleTolerance && currentScale >= maxScale && currentScale < maxScaleTolerance) {
-        return false;
-      }
-
-      // Too close to zoom-out boundary (minScale)
-      if (minScale && minScaleTolerance && currentScale > minScaleTolerance && currentScale <= minScale) {
-        return false;
-      }
-    }
-
-    return true;
+    // Redirect to overridable method
+    return this.onIsInVisibleRange(currentResolution, currentScale, effectiveScales);
   }
 
   // #endregion METHODS
